fix flow check tiket sqm & reguler
