+++ src/sqm/sqm.service.ts
@@ -11,108 +11,26 @@ export class SqmService {
     async submit_sqm(initalDto: RequestTicketDataProps, dto: TicketSQMProps) {
         const { job_id, idTelegram } = initalDto;
         const { insiden_number, speedy_number, customer_name, customer_phone, problem, description } = dto;
-        const dateNow = Date.now();
         try {
-            const find_sqm = await this.prisma.ticket_sqm.findMany({
-                where: {
-                    speedy_number
-                },
-                orderBy: {
-                    createAt: 'desc'
+            const sqm = await this.prisma.ticket_sqm.create({
+                data: {
+                    insiden_number,
+                    speedy_number,
+                    customer_name,
+                    customer_number: customer_phone,
+                    problem,
+                    description,
+                    teknisi_job_id: job_id,
+                    idTelegram
                 }
             })
+            if (sqm) return {
+                status: true,
+                statusCode: 200,
+                message: 'Create sqm successfull',
+                data: sqm
+            };
 
-            if (find_sqm.length > 0) {
-                const prevSqm = find_sqm[0];
-                const prevDate = new Date(prevSqm.createAt).getTime();
-                const countGapDate = dateNow - prevDate;
-                const day = 1000 * 60 * 60 * 24;
-                if ((countGapDate / day) < 60) {
-                    const sendData: TiketRedundantProps = {
-                        insiden_number: insiden_number,
-                        speedy_number: speedy_number,
-                        customer_name: customer_name,
-                        customer_phone: customer_phone,
-                        problem: problem,
-                        description: description,
-                        job_id: job_id,
-                        idTelegram: idTelegram,
-                    }
-                    // submit to tiket redundant
-                    const res = await this.tiket_redundant_serv.submit_tiket_redundant(sendData);
-                    if (res.statusCode == 200) {
-                        return {
-                            status: true,
-                            statusCode: 200,
-                            message: 'Create sqm successfull',
-                        }
-                    }
-                    // if (res.statusCode === 200) {
-                    //     const sqm = await this.prisma.ticket_sqm.create({
-                    //         data: {
-                    //             insiden_number,
-                    //             speedy_number,
-                    //             customer_name,
-                    //             customer_number: customer_phone,
-                    //             problem,
-                    //             description,
-                    //             teknisi_job_id: job_id,
-                    //             idTelegram
-                    //         }
-                    //     })
-                    //     if (sqm) return {
-                    //         status: true,
-                    //         statusCode: 200,
-                    //         message: 'Create sqm successfull',
-                    //         data: sqm
-                    //     };
-                    // } else {
-                    //     return {
-                    //         status: true,
-                    //         statusCode: 403,
-                    //         message: 'error submiting tiket redundant',
-                    //     }
-                    // }
-                } else {
-                    const sqm = await this.prisma.ticket_sqm.create({
-                        data: {
-                            insiden_number,
-                            speedy_number,
-                            customer_name,
-                            customer_number: customer_phone,
-                            problem,
-                            description,
-                            teknisi_job_id: job_id,
-                            idTelegram
-                        }
-                    })
-                    if (sqm) return {
-                        status: true,
-                        statusCode: 200,
-                        message: 'Create sqm successfull',
-                        data: sqm
-                    };
-                }
-            } else {
-                const sqm = await this.prisma.ticket_sqm.create({
-                    data: {
-                        insiden_number,
-                        speedy_number,
-                        customer_name,
-                        customer_number: customer_phone,
-                        problem,
-                        description,
-                        teknisi_job_id: job_id,
-                        idTelegram
-                    }
-                })
-                if (sqm) return {
-                    status: true,
-                    statusCode: 200,
-                    message: 'Create sqm successfull',
-                    data: sqm
-                };
-            }
             return {
                 status: false,
                 statusCode: 500,
